requirements.txt
django
rstr
django-model-utils